fix: move export atom fragment to fragments file
