+++ libs/frontend/modules/atom/src/Atom.fragment.graphql
@@ -20,3 +20,18 @@ fragment AtomBase on Atom {
     typeKind
   }
 }
+
+fragment Export__Atoms on Atom {
+  id
+  name
+  type
+  api {
+    id
+    name
+    # These are resolved using @ResolveField
+    typeKind
+    typeGraph {
+      ...TypeGraph
+    }
+  }
+}
