+++ libs/frontend/modules/atom/src/Atom.fragment.graphql.gen.ts
@@ -1,6 +1,8 @@
 import * as Types from '@codelab/frontend/abstract/codegen'
 
+import { TypeGraphFragment } from '../../type/src/graphql/TypeGraph.fragment.graphql.gen'
 import { gql } from '@apollo/client'
+import { TypeGraphFragmentDoc } from '../../type/src/graphql/TypeGraph.fragment.graphql.gen'
 export type AtomFragment = {
   __typename: 'Atom'
   id: string
@@ -14,6 +16,18 @@ export type AtomBaseFragment = {
   name: string
   type: Types.AtomType
   api: { id: string; name: string; typeKind: Types.TypeKind }
+}
+
+export type Export__AtomsFragment = {
+  id: string
+  name: string
+  type: Types.AtomType
+  api: {
+    id: string
+    name: string
+    typeKind: Types.TypeKind
+    typeGraph: TypeGraphFragment
+  }
 }
 
 export const AtomFragmentDoc = gql`
@@ -40,4 +54,20 @@ export const AtomBaseFragmentDoc = gql`
       typeKind
     }
   }
+`
+export const Export__AtomsFragmentDoc = gql`
+  fragment Export__Atoms on Atom {
+    id
+    name
+    type
+    api {
+      id
+      name
+      typeKind
+      typeGraph {
+        ...TypeGraph
+      }
+    }
+  }
+  ${TypeGraphFragmentDoc}
 `
