+++ libs/frontend/modules/atom/src/use-cases/atom.endpoints.graphql
@@ -38,18 +38,3 @@ mutation UpdateAtom($input: UpdateAtomInput!) {
     ...AtomBase
   }
 }
-
-fragment Export__Atoms on Atom {
-  id
-  name
-  type
-  api {
-    id
-    name
-    # These are resolved using @ResolveField
-    typeKind
-    typeGraph {
-      ...TypeGraph
-    }
-  }
-}
