+++ libs/frontend/modules/atom/src/use-cases/atom.endpoints.graphql.gen.ts
@@ -1,13 +1,16 @@
 import * as Types from '@codelab/frontend/abstract/codegen'
 
-import { AtomBaseFragment, AtomFragment } from '../Atom.fragment.graphql.gen'
-import { TypeGraphFragment } from '../../../type/src/graphql/TypeGraph.fragment.graphql.gen'
+import {
+  AtomBaseFragment,
+  Export__AtomsFragment,
+  AtomFragment,
+} from '../Atom.fragment.graphql.gen'
 import { gql } from '@apollo/client'
 import {
   AtomBaseFragmentDoc,
+  Export__AtomsFragmentDoc,
   AtomFragmentDoc,
 } from '../Atom.fragment.graphql.gen'
-import { TypeGraphFragmentDoc } from '../../../type/src/graphql/TypeGraph.fragment.graphql.gen'
 import { api, GraphqlOperationOptions } from '@codelab/frontend/model/infra/api'
 export type CreateAtomMutationVariables = Types.Exact<{
   input: Types.CreateAtomInput
@@ -59,34 +62,6 @@ export type UpdateAtomMutation = {
   updateAtom?: AtomBaseFragment | null | undefined
 }
 
-export type Export__AtomsFragment = {
-  id: string
-  name: string
-  type: Types.AtomType
-  api: {
-    id: string
-    name: string
-    typeKind: Types.TypeKind
-    typeGraph: TypeGraphFragment
-  }
-}
-
-export const Export__AtomsFragmentDoc = gql`
-  fragment Export__Atoms on Atom {
-    id
-    name
-    type
-    api {
-      id
-      name
-      typeKind
-      typeGraph {
-        ...TypeGraph
-      }
-    }
-  }
-  ${TypeGraphFragmentDoc}
-`
 export const CreateAtomGql = gql`
   mutation CreateAtom($input: CreateAtomInput!) {
     createAtom(input: $input) {
